refactor(nws): split weather client helpers
Move NWS current conditions, forecast text products, alerts, aviation helpers, hourly pressure handling, and parsers into focused modules while preserving weather_client_nws as the public compatibility surface.

Tested: NWS regression suite 90 passed; uv run ruff check src tests scripts; uv run pyright

Confidence: high
